change function signature

+++ util/util.go
@@ -3,7 +3,7 @@ package util
 import "fmt"
 
 // Gets the n-th (0-indexed) bit out of Most Significant Bit byte
-func FindBit(b byte, n uint64) bool {
+func FindBit(b byte, n int8) bool {
 	sb := byte(1 << n)
 	return sb == (sb & b)
 }
